requirements.txt
# =============================================================================
# AskVineet — Python Dependencies
# Install: pip install -r requirements.txt
# =============================================================================

# ── Agent Orchestration ───────────────────────────────────────────────────────
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38

# ── LLM Abstraction (100+ providers via one API) ─────────────────────────────
litellm==1.44.22

# ── LLM Provider SDKs (install only what you use) ────────────────────────────
openai==1.40.0
anthropic==0.34.0
google-generativeai==0.7.2

# ── Vector Stores ─────────────────────────────────────────────────────────────
chromadb==0.5.15
faiss-cpu==1.8.0.post1
langchain-chroma==0.1.4

# ── Embeddings ────────────────────────────────────────────────────────────────
sentence-transformers==3.1.1

# ── Document Loaders & OCR ───────────────────────────────────────────────────
unstructured[pdf,docx]==0.15.13
pypdf==4.3.1
python-docx==1.1.2
pytesseract==0.3.13
Pillow==10.4.0
pdf2image==1.17.0

# ── API Layer ─────────────────────────────────────────────────────────────────
fastapi==0.114.2
uvicorn[standard]==0.30.6
websockets==13.0
httpx==0.27.2
python-multipart==0.0.12

# ── Google APIs (Gmail + Calendar) ───────────────────────────────────────────
google-api-python-client==2.143.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# ── Configuration ─────────────────────────────────────────────────────────────
pydantic==2.9.2
pydantic-settings==2.5.2
pyyaml==6.0.2
python-dotenv==1.0.1

# ── Workflow & Scheduling ─────────────────────────────────────────────────────
apscheduler==3.10.4
watchdog==5.0.2

# ── UI ────────────────────────────────────────────────────────────────────────
streamlit==1.38.0

# ── Utilities ─────────────────────────────────────────────────────────────────
loguru==0.7.2
rich==13.8.1
tenacity==9.0.0
aiofiles==24.1.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0
pytest-cov==5.0.0
jsonschema==4.23.0
